lxd/storage/backend/lxd: Updates CreateInstanceFromMigration to handle remote shared storage cluster moves

+++ lxd/storage/backend_lxd.go
@@ -1805,62 +1805,70 @@ func (b *lxdBackend) CreateInstanceFromMigration(inst instance.Instance, conn io
 	// detecting if the instance DB record exists or not. If we get here then something has gone wrong.
 	if args.Refresh && !volExists {
 		return fmt.Errorf("Cannot refresh volume, doesn't exist on migration target storage")
-	} else if !args.Refresh && volExists {
-		return fmt.Errorf("Cannot create volume, already exists on migration target storage")
 	}
 
 	revert := revert.New()
 	defer revert.Fail()
 
-	if !args.Refresh {
-		// Validate config and create database entry for new storage volume if not refreshing.
-		// Strip unsupported config keys (in case the export was made from a different type of storage pool).
-		err = VolumeDBCreate(b, inst.Project().Name, inst.Name(), volumeDescription, volType, false, volumeConfig, inst.CreationDate(), time.Time{}, contentType, true, true)
-		if err != nil {
-			return err
-		}
+	isRemoteClusterMove := args.ClusterMoveSourceName != "" && b.driver.Info().Remote
 
-		revert.Add(func() { _ = VolumeDBDelete(b, inst.Project().Name, inst.Name(), volType) })
+	if !args.Refresh {
+		if volExists {
+			if !isRemoteClusterMove {
+				return fmt.Errorf("Cannot create volume, already exists on migration target storage")
+			}
+		} else {
+			// Validate config and create database entry for new storage volume if not refreshing.
+			// Strip unsupported config keys (in case the export was made from a different type of storage pool).
+			err = VolumeDBCreate(b, inst.Project().Name, inst.Name(), volumeDescription, volType, false, volumeConfig, inst.CreationDate(), time.Time{}, contentType, true, true)
+			if err != nil {
+				return err
+			}
+
+			revert.Add(func() { _ = VolumeDBDelete(b, inst.Project().Name, inst.Name(), volType) })
+		}
 	}
 
-	for i, snapName := range args.Snapshots {
-		newSnapshotName := drivers.GetSnapshotVolumeName(inst.Name(), snapName)
-		snapConfig := volumeConfig           // Use parent volume config by default.
-		snapDescription := volumeDescription // Use parent volume description by default.
-		snapExpiryDate := time.Time{}
-		snapCreationDate := time.Time{}
+	if !isRemoteClusterMove {
+		for i, snapName := range args.Snapshots {
+			newSnapshotName := drivers.GetSnapshotVolumeName(inst.Name(), snapName)
+			snapConfig := volumeConfig           // Use parent volume config by default.
+			snapDescription := volumeDescription // Use parent volume description by default.
+			snapExpiryDate := time.Time{}
+			snapCreationDate := time.Time{}
 
-		// If the source snapshot config is available, use that.
-		if srcInfo != nil && srcInfo.Config != nil {
-			if len(srcInfo.Config.Snapshots) >= i-1 && srcInfo.Config.Snapshots[i] != nil && srcInfo.Config.Snapshots[i].Name == snapName {
-				// Use instance snapshot's creation date if snap info available.
-				snapCreationDate = srcInfo.Config.Snapshots[i].CreatedAt
-			}
-
-			if len(srcInfo.Config.VolumeSnapshots) >= i-1 && srcInfo.Config.VolumeSnapshots[i] != nil && srcInfo.Config.VolumeSnapshots[i].Name == snapName {
-				// Check if snapshot volume config is available then use it.
-				snapDescription = srcInfo.Config.VolumeSnapshots[i].Description
-				snapConfig = srcInfo.Config.VolumeSnapshots[i].Config
-
-				if srcInfo.Config.VolumeSnapshots[i].ExpiresAt != nil {
-					snapExpiryDate = *srcInfo.Config.VolumeSnapshots[i].ExpiresAt
+			// If the source snapshot config is available, use that.
+			if srcInfo != nil && srcInfo.Config != nil {
+				if len(srcInfo.Config.Snapshots) >= i-1 && srcInfo.Config.Snapshots[i] != nil && srcInfo.Config.Snapshots[i].Name == snapName {
+					// Use instance snapshot's creation date if snap info available.
+					snapCreationDate = srcInfo.Config.Snapshots[i].CreatedAt
 				}
 
-				// Use volume's creation date if available.
-				if !srcInfo.Config.VolumeSnapshots[i].CreatedAt.IsZero() {
-					snapCreationDate = srcInfo.Config.VolumeSnapshots[i].CreatedAt
+				if len(srcInfo.Config.VolumeSnapshots) >= i-1 && srcInfo.Config.VolumeSnapshots[i] != nil && srcInfo.Config.VolumeSnapshots[i].Name == snapName {
+					// Check if snapshot volume config is available then use it.
+					snapDescription = srcInfo.Config.VolumeSnapshots[i].Description
+					snapConfig = srcInfo.Config.VolumeSnapshots[i].Config
+
+					if srcInfo.Config.VolumeSnapshots[i].ExpiresAt != nil {
+						snapExpiryDate = *srcInfo.Config.VolumeSnapshots[i].ExpiresAt
+					}
+
+					// Use volume's creation date if available.
+					if !srcInfo.Config.VolumeSnapshots[i].CreatedAt.IsZero() {
+						snapCreationDate = srcInfo.Config.VolumeSnapshots[i].CreatedAt
+					}
 				}
 			}
-		}
 
-		// Validate config and create database entry for new storage volume.
-		// Strip unsupported config keys (in case the export was made from a different type of storage pool).
-		err = VolumeDBCreate(b, inst.Project().Name, newSnapshotName, snapDescription, volType, true, snapConfig, snapCreationDate, snapExpiryDate, contentType, true, true)
-		if err != nil {
-			return err
-		}
+			// Validate config and create database entry for new storage volume.
+			// Strip unsupported config keys (in case the export was made from a different type of storage pool).
+			err = VolumeDBCreate(b, inst.Project().Name, newSnapshotName, snapDescription, volType, true, snapConfig, snapCreationDate, snapExpiryDate, contentType, true, true)
+			if err != nil {
+				return err
+			}
 
-		revert.Add(func() { _ = VolumeDBDelete(b, inst.Project().Name, newSnapshotName, volType) })
+			revert.Add(func() { _ = VolumeDBDelete(b, inst.Project().Name, newSnapshotName, volType) })
+		}
 	}
 
 	// Generate the effective root device volume for instance.
@@ -1887,7 +1895,7 @@ func (b *lxdBackend) CreateInstanceFromMigration(inst instance.Instance, conn io
 
 	var preFiller drivers.VolumeFiller
 
-	if !args.Refresh {
+	if !args.Refresh && !isRemoteClusterMove {
 		// If the negotiated migration method is rsync and the instance's base image is
 		// already on the host then setup a pre-filler that will unpack the local image
 		// to try and speed up the rsync of the incoming volume by avoiding the need to
@@ -1933,7 +1941,9 @@ func (b *lxdBackend) CreateInstanceFromMigration(inst instance.Instance, conn io
 		return err
 	}
 
-	revert.Add(func() { _ = b.DeleteInstance(inst, op) })
+	if !isRemoteClusterMove {
+		revert.Add(func() { _ = b.DeleteInstance(inst, op) })
+	}
 
 	err = b.ensureInstanceSymlink(inst.Type(), inst.Project().Name, inst.Name(), vol.MountPath())
 	if err != nil {
